dev-requirements.txt
#
# Copyright (C) 2023 Intel Corporation
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

flake8==7.0.0
black==22.3.0
mypy==0.942
isort==5.8.0
pytest==6.2.5
requests==2.27.1
boto3-stubs==1.26.13
types-requests==2.28.3
types-jsonschema==4.17.0.1
stac-fastapi.types==2.4.3
types-cachetools==5.2.1
types-python-dateutil==2.8.19.4
types-PyYAML==6.0.12.2
types-simplejson==3.18.0.0
types-setuptools==65.6.0.1
types-certifi==2021.10.8.3
types-aiofiles==22.1.0.4
types-pycurl==7.45.1.4
types-orjson==3.6.2
